fix(communities): publish community images that were uploaded private
Digest emails group unread discussions by community and render GP Team.image
as <img src> with an absolute URL, so a private one 403s in a mail client the
same way an avatar does. Making new uploads public only helps the next upload;
the images set while frappe-ui defaulted uploads to private are broken today.

Same shape as the avatar patch, so the machinery is now shared in
gameplan/utils/file_privacy.py: save the File so handle_is_private_changed
moves the blob and rewrites file_url, repoint the referencing field, recover
a run that stopped between the two, and skip anything already public.

Kept as two patch entries rather than one parameterised patch. patches.txt is
what frappe tracks, so a site that has already logged the avatar patch would
never run the community half if they shared an entry.

GP Team.image is attached on save, being an Attach Image on a doctype that is
not a child table, so frappe can resolve the referencing field on its own and
the patch's repoint is belt and braces here. It is load bearing wherever the
file is unattached. A test covers both, and removing the repoint fails the
unattached one while the attached one still passes.

Renamed test_avatar_visibility to test_public_image_files: it now covers both
patches, and the blob teardown they share is subtle enough to want one copy.
